# Use bulk string writes for text formats