--- a/prometheus-metrics-exposition-textformats/src/main/java/io/prometheus/metrics/expositionformats/TextFormatUtil.java
+++ b/prometheus-metrics-exposition-textformats/src/main/java/io/prometheus/metrics/expositionformats/TextFormatUtil.java
@@ -35,21 +35,53 @@ public class TextFormatUtil {
   }
 
   static void writeEscapedLabelValue(Writer writer, String s) throws IOException {
-    for (int i = 0; i < s.length(); i++) {
-      char c = s.charAt(i);
+    // optimize for the common case where no escaping is needed
+    int start = 0;
+    // #indexOf is a vectorized intrinsic
+    int backslashIndex = s.indexOf('\\', start);
+    int quoteIndex = s.indexOf('\"', start);
+    int newlineIndex = s.indexOf('\n', start);
+
+    int allEscapesIndex = backslashIndex & quoteIndex & newlineIndex;
+    while (allEscapesIndex != -1) {
+      int escapeStart = Integer.MAX_VALUE;
+      if (backslashIndex != -1) {
+        escapeStart = backslashIndex;
+      }
+      if (quoteIndex != -1) {
+        escapeStart = Math.min(escapeStart, quoteIndex);
+      }
+      if (newlineIndex != -1) {
+        escapeStart = Math.min(escapeStart, newlineIndex);
+      }
+
+      // bulk write up to the first character that needs to be escaped
+      if (escapeStart > start) {
+        writer.write(s, start, escapeStart - start);
+      }
+      char c = s.charAt(escapeStart);
+      start = escapeStart + 1;
       switch (c) {
         case '\\':
-          writer.append("\\\\");
+          writer.write("\\\\");
+          backslashIndex = s.indexOf('\\', start);
           break;
         case '\"':
-          writer.append("\\\"");
+          writer.write("\\\"");
+          quoteIndex = s.indexOf('\"', start);
           break;
         case '\n':
-          writer.append("\\n");
+          writer.write("\\n");
+          newlineIndex = s.indexOf('\n', start);
           break;
-        default:
-          writer.append(c);
       }
+
+      allEscapesIndex = backslashIndex & quoteIndex & newlineIndex;
+    }
+    // up until the end nothing needs to be escaped anymore
+    int remaining = s.length() - start;
+    if (remaining > 0) {
+      writer.write(s, start, remaining);
     }
   }