@zeitlinger 2025-02-18
> The build fails because of the lack of test coverage in `prometheus-metrics-exposition-textformats`. Should I increase the test coverage or lower the exception in pom.xml?

I merged a PR to increase the coverage